+++ demo/public/samples/go.go
@@ -1,0 +1,8 @@
+package main
+
+import "fmt"
+
+func greet(name string) string {
+  // comment
+  return fmt.Sprintf("Hello, %s!", name)
+}